Move kubernetes specific sources to platform/kube
