+++ platform/kube/consumer.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package env
+package kube
 
 // ConfigMapConsumer relies on an agent and FS watch mechanism on the mounted
 // volume to kick-off a restart on the config update.
